server: Implement journal object load/save
The frontend server can now load and save journal objects
(which are really just plain JSON structures).

As of now, the backend does not manipulate the data, thus we do
not need to have an exact model of the JSON scheme, but just store
the raw object as-is in MongoDB.

A journal must have a unique pincode assigned to it so the frontend
is able to request journals by (pincode) parameter.

body-parser is added as dependency to be able to parse JSON from
HTTP request bodies.

Signed-off-by: Christopher N. Hesse <raymanfx@gmail.com>