server: controller: Restructure pin create flow
Preparing for the upcoming unit test integration: chained
mongoose queries (promises) are bad because a single 'await'
only waits for the first one to resolve, apparently.
This is an issue because in unit tests, queries *must* be
carried out in a synchronous fashion, otherwise they will always
fail because of pending promises.

Signed-off-by: Christopher N. Hesse <raymanfx@gmail.com>