server: Implement result validation code
The actual code generation is an instance function for each MongoDB Result
document instance.

Note that to facilitate unit testing, the result controller test had to be
modified to create a 'real' model instance in the test data set (because the
previous JSON ResultDocs array obviously did not support the model instance
method).

All unit tests should be moved to this approach in the future.

Signed-off-by: Christopher N. Hesse <raymanfx@gmail.com>