app: Provide set-specific evaluation texts
Every set needs to have a score independent evaluation
text which shows general evaluation-information.
In addition to that: A set needs to provide information
dependent on the users score.
It was neccessary to extend the set definition by
another attribute:

"sets": [
    {
      ...
      "evaluationTexts": {
        "scoreIndependent": "?ref{3001-1}",
        "scoreDependent": [
          [33, "?ref{3001-2}"],
          [66, "?ref{3001-3}"],
          [100, "?ref{3001-4}"]
        ]
      }
    },...

The 'scoreDependent' array contains a tuple consisting of
two elements:
    - number: the percentage
    - string: the score specific text

If the users score is less than or equal to the defined
number, the text is shown. This way it is possible
to define the range very fine-grained or very broadly.