app: Change LocalStorageService and delete guards
Every call to the localStorage object (except for the ones in the
resource service) are executed like the following:

For retrieving objects:
const pin = storageInstance.retrieveStorage(StorageItem.PIN);

For storing objects:
storageInstance.persistInStorage(StorageItem.PIN, pin);

If at any point an object vital to the usage of the application is
missing from the local storage (user deleted cache) the router routes
back to the dashboard. This renders the guards obsolete.

Also: Minor changes in the resource service and the start-test comp to
support slow connections.