app: Allow every test to be a speedtest
As of now every test can be a speedtest by setting the 'seconds'
attribute to the specific test in the config file.
Every journal is now initalized with 'null' values instead of false
values to prevent users from taking the same speedtest twice.
TODO: The original speedtest category is still named speed, it should be
renamed to match.