server: user: Improve validation code
The modulo operator '%' can now be used to compute a checksum
value of a capture group.

Previously, this was a 'dummy' feature and would only calculate
a value for the very last element.

The checksum is computed as sum of Unicode character values of
all elements in the capture group.

Signed-off-by: Christopher N. Hesse <raymanfx@gmail.com>