app: Simplify image integration in config file
It is no longer neccessary to include the backend api url along with the
image. Just do the following inside the course config file.

"description": "This is my description with an image: ``image.jpg``."

If you want an image were the text flows arround on the left/right side,
wrap the image inside the following div:

"description": "This is my description with an image:
<div class='left-img'>``image.jpg``</div>."