app: bug fix: failed to load evaluation
Bug: User starts test-procedure, closes browser (cache is gone).
Retrieves his progress via pin-input, gets an error when loading the
evaluation component because the course object in the local storage
can not be found.
Fix: Add course object to local storage, when receiving the minimal
journal structure object from the backend, because the proper journal
structure does not have a reference to its respective course.